HAWQ-1490. Refactored Dummy profile into pxf-api project
