+++ pxf/build.gradle
@@ -482,13 +482,6 @@ project('pxf-jdbc') {
     }
 }
 
-project('pxf-local') {
-    dependencies {
-        compile(project(':pxf-api'))
-        compile(project(':pxf-service'))
-    }
-}
-
 def buildNumber() {
     System.getenv('BUILD_NUMBER') ?: System.getProperty('user.name')
 }
@@ -497,7 +490,7 @@ task wrapper(type: Wrapper) {
     gradleVersion = '2.11'
 }
 
-def distSubprojects = subprojects - project(':pxf-api')
+def distSubprojects = subprojects
 
 task release(type: Copy, dependsOn: [subprojects.build, subprojects.javadoc, distSubprojects.buildRpm, distSubprojects.distTar]) {
     delete 'build'
