+++ pxf/settings.gradle
@@ -27,4 +27,3 @@ include 'pxf-hive'
 include 'pxf-hbase'
 include 'pxf-json'
 include 'pxf-jdbc'
-include 'pxf-local'
